Rename ddd-core package
