+++ account/src/main/kotlin/htnk128/kotlin/ddd/sample/account/adapter/gateway/messaging/AccountEventSpringPublisher.kt
@@ -1,7 +1,7 @@
 package htnk128.kotlin.ddd.sample.account.adapter.gateway.messaging
 
 import htnk128.kotlin.ddd.sample.account.domain.model.account.AccountEvent
-import htnk128.kotlin.ddd.sample.dddcore.domain.DomainEventPublisher
+import htnk128.kotlin.ddd.sample.ddd.core.domain.DomainEventPublisher
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
