+++ account/src/main/kotlin/htnk128/kotlin/ddd/sample/account/domain/model/account/NamePronunciation.kt
@@ -1,6 +1,6 @@
 package htnk128.kotlin.ddd.sample.account.domain.model.account
 
-import htnk128.kotlin.ddd.sample.dddcore.domain.SomeValueObject
+import htnk128.kotlin.ddd.sample.ddd.core.domain.SomeValueObject
 
 /**
  * アカウントの発音を表現する。
