+++ account/src/main/kotlin/htnk128/kotlin/ddd/sample/account/domain/model/addressbook/AccountAddress.kt
@@ -1,6 +1,6 @@
 package htnk128.kotlin.ddd.sample.account.domain.model.addressbook
 
-import htnk128.kotlin.ddd.sample.dddcore.domain.ValueObject
+import htnk128.kotlin.ddd.sample.ddd.core.domain.ValueObject
 import java.time.Instant
 
 /**
