+++ ddd-core/src/main/kotlin/htnk128/kotlin/ddd/sample/ddd/core/domain/DomainEventPublisher.kt
@@ -1,4 +1,4 @@
-package htnk128.kotlin.ddd.sample.dddcore.domain
+package htnk128.kotlin.ddd.sample.ddd.core.domain
 
 /**
  * [DomainEvent]の出版インターフェース。
