+++ ddd-core/src/main/kotlin/htnk128/kotlin/ddd/sample/ddd/core/domain/Entity.kt
@@ -1,4 +1,4 @@
-package htnk128.kotlin.ddd.sample.dddcore.domain
+package htnk128.kotlin.ddd.sample.ddd.core.domain
 
 /**
  * DDDにおけるエンティティの概念。
