+++ ddd-core/src/main/kotlin/htnk128/kotlin/ddd/sample/ddd/core/domain/SomeIdentity.kt
@@ -1,4 +1,4 @@
-package htnk128.kotlin.ddd.sample.dddcore.domain
+package htnk128.kotlin.ddd.sample.ddd.core.domain
 
 /**
  * 何らかのドメインを識別するIDを表現した値オブジェクト。
